third_party/WebKit/LayoutTests/crypto/exportKey-badParameters-expected.txt

CONSOLE ERROR: Invalid keyFormat argument
CONSOLE ERROR: Invalid keyFormat argument
CONSOLE ERROR: Invalid keyFormat argument
Tests exportKey() given bad inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.exportKey('raw', null) threw exception TypeError: Failed to execute 'exportKey' on 'SubtleCrypto': Invalid key argument.
PASS crypto.subtle.exportKey('raw', 3) threw exception TypeError: Failed to execute 'exportKey' on 'SubtleCrypto': Invalid key argument.
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
